+++ src/main/java/org/embulk/input/marketo/delegate/MarketoBaseInputPluginDelegate.java
@@ -1,7 +1,9 @@
 package org.embulk.input.marketo.delegate;
 
+import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.annotations.VisibleForTesting;
-
+import com.google.common.collect.Lists;
+import org.apache.commons.lang3.StringUtils;
 import org.embulk.base.restclient.DefaultServiceDataSplitter;
 import org.embulk.base.restclient.RestClientInputPluginDelegate;
 import org.embulk.base.restclient.RestClientInputTaskBase;
@@ -11,6 +13,7 @@ import org.embulk.base.restclient.record.ServiceRecord;
 import org.embulk.config.Config;
 import org.embulk.config.ConfigDefault;
 import org.embulk.config.ConfigDiff;
+import org.embulk.config.ConfigException;
 import org.embulk.config.TaskReport;
 import org.embulk.input.marketo.MarketoService;
 import org.embulk.input.marketo.MarketoServiceImpl;
@@ -19,16 +22,26 @@ import org.embulk.spi.Exec;
 import org.embulk.spi.PageBuilder;
 import org.embulk.spi.Schema;
 import org.joda.time.DateTime;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Set;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 /**
  * Created by tai.khuu on 9/18/17.
  */
 public abstract class MarketoBaseInputPluginDelegate<T extends MarketoBaseInputPluginDelegate.PluginTask> implements RestClientInputPluginDelegate<T>
 {
+    private final Logger logger = LoggerFactory.getLogger(getClass());
     public static final int PREVIEW_RECORD_LIMIT = 15;
+    public static final String ID_LIST_SEPARATOR_CHAR = ",";
+
     public interface PluginTask
             extends RestClientInputTaskBase, MarketoRestClient.PluginTask
     {
@@ -88,5 +101,58 @@ public abstract class MarketoBaseInputPluginDelegate<T extends MarketoBaseInputP
     public ServiceDataSplitter<T> buildServiceDataSplitter(T task)
     {
         return new DefaultServiceDataSplitter();
+    }
+
+    protected Iterable<ObjectNode> getObjectsByIds(String[] inputIds, Function<Set<String>, Iterable<ObjectNode>> getByIdFunction)
+    {
+        final Set<String> ids = new HashSet<>();
+        final List<String> invalidIds = new ArrayList<>();
+
+        for (int i = 0; i < inputIds.length; i++) {
+            String currentId = StringUtils.trimToNull(inputIds[i]);
+            // ignore null or empty ids
+            if (currentId == null) {
+                continue;
+            }
+
+            // ignore or throw for NaN ids
+            if (StringUtils.isNumeric(currentId)) {
+                ids.add(currentId);
+            }
+            else {
+                invalidIds.add(inputIds[i]);
+            }
+        }
+
+        if (ids.isEmpty()) {
+            throw new ConfigException("No valid Id specified");
+        }
+
+        if (!invalidIds.isEmpty()) {
+            logger.warn("Ignore invalid Id(s): {}", invalidIds);
+        }
+
+        List<ObjectNode> actualList = Lists.newArrayList(getByIdFunction.apply(ids));
+        if (actualList.isEmpty()) {
+            throw new ConfigException("No valid Id found");
+        }
+
+        if (actualList.size() != ids.size()) {
+            logNoneExistedIds(ids, actualList);
+        }
+
+        return actualList;
+    }
+
+    private void logNoneExistedIds(Set<String> ids, List<ObjectNode> actualList)
+    {
+        List<String> actualIds = actualList.parallelStream().map(n -> String.valueOf(n.get("id").asInt())).collect(Collectors.toList());
+        List<String> missingIds = new ArrayList<>();
+        for (String id : ids) {
+            if (!actualIds.contains(id)) {
+                missingIds.add(id);
+            }
+        }
+        logger.warn("Ignore not exists Id(s): {}", missingIds);
     }
 }