+++ src/test/java/org/embulk/input/marketo/delegate/CustomObjectInputPluginTest.java
@@ -3,7 +3,6 @@ package org.embulk.input.marketo.delegate;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.google.common.base.Optional;
 import org.embulk.EmbulkTestRuntime;
 import org.embulk.base.restclient.ServiceResponseMapper;
 import org.embulk.base.restclient.record.RecordImporter;
@@ -16,6 +15,7 @@ import org.embulk.input.marketo.rest.MarketoRestClient;
 import org.embulk.input.marketo.rest.RecordPagingIterable;
 import org.embulk.spi.PageBuilder;
 import org.embulk.spi.Schema;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -50,33 +50,118 @@ public class CustomObjectInputPluginTest
         Mockito.doReturn(mockMarketoRestClient).when(customObjectInputPlugin).createMarketoRestClient(Mockito.any(CustomObjectInputPlugin.PluginTask.class));
     }
 
-    @Test(expected = ConfigException.class)
-    public void validateCustomObjectFilterTypeError()
+    @Test
+    public void testValidPluginTask()
     {
-        CustomObjectInputPlugin.PluginTask pluginTask = Mockito.mock(CustomObjectInputPlugin.PluginTask.class);
-        Mockito.when(pluginTask.getCustomObjectFilterType()).thenReturn("");
-        customObjectInputPlugin.validateInputTask(pluginTask);
-    }
-
-    @Test(expected = ConfigException.class)
-    public void validateCustomObjectAPINameError()
-    {
-        CustomObjectInputPlugin.PluginTask pluginTask = Mockito.mock(CustomObjectInputPlugin.PluginTask.class);
-        Mockito.when(pluginTask.getCustomObjectAPIName()).thenReturn("");
-        customObjectInputPlugin.validateInputTask(pluginTask);
-    }
-
-    @Test(expected = ConfigException.class)
-    public void validateFromValueGreaterThanToValueError()
-    {
-        CustomObjectInputPlugin.PluginTask pluginTask = Mockito.mock(CustomObjectInputPlugin.PluginTask.class);
-        Mockito.when(pluginTask.getFromValue()).thenReturn(100);
-        Mockito.when(pluginTask.getToValue()).thenReturn(Optional.of(90));
+        CustomObjectInputPlugin.PluginTask pluginTask = configSource.loadConfig(CustomObjectInputPlugin.PluginTask.class);
+        // should not cause exception
         customObjectInputPlugin.validateInputTask(pluginTask);
     }
 
     @Test
-    public void testRun() throws IOException
+    public void testCustomObjectFilterTypeError()
+    {
+        CustomObjectInputPlugin.PluginTask pluginTask = configSource
+                .set("custom_object_filter_type", "")
+                .loadConfig(CustomObjectInputPlugin.PluginTask.class);
+        Assert.assertThrows(ConfigException.class, () -> customObjectInputPlugin.validateInputTask(pluginTask));
+    }
+
+    @Test
+    public void testCustomObjectAPINameError()
+    {
+        CustomObjectInputPlugin.PluginTask pluginTask = configSource
+                .set("custom_object_api_name", "")
+                .loadConfig(CustomObjectInputPlugin.PluginTask.class);
+        Assert.assertThrows(ConfigException.class, () -> customObjectInputPlugin.validateInputTask(pluginTask));
+    }
+
+    @Test
+    public void testFromValueGreaterThanToValueError()
+    {
+        CustomObjectInputPlugin.PluginTask pluginTask = configSource
+                .set("custom_object_filter_from_value", "10")
+                .set("custom_object_filter_to_value", "5").loadConfig(CustomObjectInputPlugin.PluginTask.class);
+        Assert.assertThrows(ConfigException.class, () -> customObjectInputPlugin.validateInputTask(pluginTask));
+    }
+
+    @Test
+    public void testEmptyStringFilterValues()
+    {
+        CustomObjectInputPlugin.PluginTask pluginTask = configSource
+                .set("custom_object_filter_values", "").loadConfig(CustomObjectInputPlugin.PluginTask.class);
+
+        Assert.assertThrows(ConfigException.class, () -> customObjectInputPlugin.validateInputTask(pluginTask));
+    }
+
+    @Test
+    public void testAllEmptyStringFilterValues()
+    {
+        CustomObjectInputPlugin.PluginTask pluginTask = configSource
+                .set("custom_object_filter_values", ",, , ").loadConfig(CustomObjectInputPlugin.PluginTask.class);
+
+        Assert.assertThrows(ConfigException.class, () -> customObjectInputPlugin.validateInputTask(pluginTask));
+    }
+
+    @Test
+    public void testRunStringFilterValues() throws IOException
+    {
+        RecordPagingIterable<ObjectNode> mockRecordPagingIterable = mockAndGetResponse();
+        Mockito.when(mockMarketoRestClient.getCustomObject(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyString())).thenReturn(mockRecordPagingIterable);
+
+        CustomObjectInputPlugin.PluginTask task = configSource.set("custom_object_filter_values", "value1,value2,value3,value4").loadConfig(CustomObjectInputPlugin.PluginTask.class);
+        ServiceResponseMapper<? extends ValueLocator> mapper = customObjectInputPlugin.buildServiceResponseMapper(task);
+        RecordImporter recordImporter = mapper.createRecordImporter();
+        PageBuilder mockPageBuilder = Mockito.mock(PageBuilder.class);
+        customObjectInputPlugin.ingestServiceData(task, recordImporter, 1, mockPageBuilder);
+        Mockito.verify(mockMarketoRestClient, Mockito.times(1)).getCustomObject(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyString());
+
+        verifyAfterRun(mapper, mockPageBuilder);
+    }
+
+    @Test
+    public void testRunWithFilterRange() throws IOException
+    {
+        RecordPagingIterable<ObjectNode> mockRecordPagingIterable = mockAndGetResponse();
+        Mockito.when(mockMarketoRestClient.getCustomObject(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyString())).thenReturn(mockRecordPagingIterable);
+
+        CustomObjectInputPlugin.PluginTask task = configSource.loadConfig(CustomObjectInputPlugin.PluginTask.class);
+        ServiceResponseMapper<? extends ValueLocator> mapper = customObjectInputPlugin.buildServiceResponseMapper(task);
+        RecordImporter recordImporter = mapper.createRecordImporter();
+        PageBuilder mockPageBuilder = Mockito.mock(PageBuilder.class);
+        customObjectInputPlugin.ingestServiceData(task, recordImporter, 1, mockPageBuilder);
+        Mockito.verify(mockMarketoRestClient, Mockito.times(1)).getCustomObject(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyString());
+
+        verifyAfterRun(mapper, mockPageBuilder);
+    }
+
+    @Test
+    public void testRunFromOnlyFilter() throws IOException
+    {
+        RecordPagingIterable<ObjectNode> mockRecordPagingIterable = mockAndGetResponse();
+        Mockito.when(mockMarketoRestClient.getCustomObject(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyInt(), Mockito.isNull())).thenReturn(mockRecordPagingIterable);
+
+        CustomObjectInputPlugin.PluginTask task = configSource.remove("custom_object_filter_to_value").loadConfig(CustomObjectInputPlugin.PluginTask.class);
+        ServiceResponseMapper<? extends ValueLocator> mapper = customObjectInputPlugin.buildServiceResponseMapper(task);
+        RecordImporter recordImporter = mapper.createRecordImporter();
+        PageBuilder mockPageBuilder = Mockito.mock(PageBuilder.class);
+        customObjectInputPlugin.ingestServiceData(task, recordImporter, 1, mockPageBuilder);
+        Mockito.verify(mockMarketoRestClient, Mockito.times(1)).getCustomObject(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyInt(), Mockito.isNull());
+
+        verifyAfterRun(mapper, mockPageBuilder);
+    }
+
+    private void verifyAfterRun(ServiceResponseMapper<? extends ValueLocator> mapper, PageBuilder mockPageBuilder)
+    {
+        Mockito.verify(mockMarketoRestClient, Mockito.times(1)).describeCustomObject(Mockito.anyString());
+        Schema schema = mapper.getEmbulkSchema();
+        ArgumentCaptor<Long> longArgumentCaptor = ArgumentCaptor.forClass(Long.class);
+        Mockito.verify(mockPageBuilder, Mockito.times(3)).setLong(Mockito.eq(schema.lookupColumn("mk_id")), longArgumentCaptor.capture());
+        List<Long> allValues = longArgumentCaptor.getAllValues();
+        assertArrayEquals(new Long[]{1L, 2L, 3L}, allValues.toArray());
+    }
+
+    private RecordPagingIterable<ObjectNode> mockAndGetResponse() throws IOException
     {
         RecordPagingIterable<ObjectNode> mockRecordPagingIterable = Mockito.mock(RecordPagingIterable.class);
         JavaType javaType = OBJECT_MAPPER.getTypeFactory().constructParametrizedType(List.class, List.class, ObjectNode.class);
@@ -84,19 +169,7 @@ public class CustomObjectInputPluginTest
         JavaType marketoFieldsType = OBJECT_MAPPER.getTypeFactory().constructParametrizedType(List.class, List.class, MarketoField.class);
         List<MarketoField> marketoFields = OBJECT_MAPPER.readValue(this.getClass().getResourceAsStream("/fixtures/custom_object_describe_marketo_fields_full.json"), marketoFieldsType);
         Mockito.when(mockRecordPagingIterable.iterator()).thenReturn(objectNodeList.iterator());
-        Mockito.when(mockMarketoRestClient.getCustomObject(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyInt(), Mockito.anyInt())).thenReturn(mockRecordPagingIterable);
         Mockito.when(mockMarketoRestClient.describeCustomObject(Mockito.anyString())).thenReturn(marketoFields);
-        CustomObjectInputPlugin.PluginTask task = configSource.loadConfig(CustomObjectInputPlugin.PluginTask.class);
-        ServiceResponseMapper<? extends ValueLocator> mapper = customObjectInputPlugin.buildServiceResponseMapper(task);
-        RecordImporter recordImporter = mapper.createRecordImporter();
-        PageBuilder mockPageBuilder = Mockito.mock(PageBuilder.class);
-        customObjectInputPlugin.ingestServiceData(task, recordImporter, 1, mockPageBuilder);
-        Mockito.verify(mockMarketoRestClient, Mockito.times(1)).getCustomObject(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyInt(), Mockito.anyInt());
-        Mockito.verify(mockMarketoRestClient, Mockito.times(1)).describeCustomObject(Mockito.anyString());
-        Schema embulkSchema = mapper.getEmbulkSchema();
-        ArgumentCaptor<Long> longArgumentCaptor = ArgumentCaptor.forClass(Long.class);
-        Mockito.verify(mockPageBuilder, Mockito.times(3)).setLong(Mockito.eq(embulkSchema.lookupColumn("mk_id")), longArgumentCaptor.capture());
-        List<Long> allValues = longArgumentCaptor.getAllValues();
-        assertArrayEquals(new Long[]{1L, 2L, 3L}, allValues.toArray());
+        return mockRecordPagingIterable;
     }
 }