+++ src/test/java/org/embulk/input/marketo/delegate/LeadWithListInputPluginTest.java
@@ -10,6 +10,7 @@ import org.embulk.base.restclient.record.RecordImporter;
 import org.embulk.base.restclient.record.ValueLocator;
 import org.embulk.config.ConfigLoader;
 import org.embulk.config.ConfigSource;
+import org.embulk.input.marketo.MarketoService;
 import org.embulk.input.marketo.MarketoUtils;
 import org.embulk.input.marketo.model.MarketoField;
 import org.embulk.input.marketo.rest.MarketoRestClient;
@@ -24,10 +25,13 @@ import org.mockito.Mockito;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyList;
+import static org.mockito.ArgumentMatchers.anySet;
 import static org.mockito.ArgumentMatchers.eq;
 
 /**
@@ -97,5 +101,32 @@ public class LeadWithListInputPluginTest
         List<Long> allValues = longArgumentCaptor.getAllValues();
         long actualValue = allValues.get(0);
         assertEquals(103280L, actualValue);
+    }
+
+    @Test
+    public void testGetListsByIds()
+    {
+        ConfigSource cfgWithInputIds = configSource.set("list_ids", "123,12333").set("skip_invalid_list_id", true);
+        LeadWithListInputPlugin.PluginTask task = cfgWithInputIds.loadConfig(LeadWithListInputPlugin.PluginTask.class);
+        MarketoService service = Mockito.mock(MarketoService.class);
+        Mockito.doReturn(Arrays.asList(new ObjectMapper().createObjectNode().put("id", "123"))).when(service).getListsByIds(anySet());
+
+        leadWithListInputPlugin.getServiceRecords(service, task);
+
+        Mockito.verify(service, Mockito.times(1)).getListsByIds(anySet());
+        Mockito.verify(service, Mockito.never()).getLists();
+        Mockito.verify(service, Mockito.times(1)).getAllListLead(anyList(), anyList());
+    }
+
+    @Test
+    public void testGetAllLists()
+    {
+        LeadWithListInputPlugin.PluginTask task = configSource.loadConfig(LeadWithListInputPlugin.PluginTask.class);
+        MarketoService service = Mockito.mock(MarketoService.class);
+        leadWithListInputPlugin.getServiceRecords(service, task);
+
+        Mockito.verify(service, Mockito.never()).getListsByIds(anySet());
+        Mockito.verify(service, Mockito.times(1)).getLists();
+        Mockito.verify(service, Mockito.times(1)).getAllListLead(anyList(), anyList());
     }
 }